[AUD-1280] Move models in src to common
