+++ src/components/nav/desktop/NavHeader.js
@@ -5,13 +5,13 @@ import PropTypes from 'prop-types'
 
 import { ReactComponent as AudiusLogoHorizontal } from 'assets/img/audiusLogoHorizontal.svg'
 import { ReactComponent as IconNotification } from 'assets/img/iconNotification.svg'
+import Theme from 'common/models/Theme'
 import { StringKeys } from 'common/services/remote-config'
 import { formatCount } from 'common/utils/formatUtil'
 import NavButton from 'components/nav/desktop/NavButton'
 import NavPopupMenu from 'components/nav/desktop/NavPopupMenu'
 import NotificationPanel from 'components/notification/NotificationPanel'
 import { useRemoteVar } from 'hooks/useRemoteConfig'
-import Theme from 'models/Theme'
 import { HOME_PAGE, BASE_URL, stripBaseUrl } from 'utils/route'
 import { getTheme } from 'utils/theme/theme'
