+++ src/pages/artist-dashboard-page/ArtistDashboardPage.tsx
@@ -11,6 +11,7 @@ import { Dispatch } from 'redux'
 
 import { ID } from 'common/models/Identifiers'
 import Status from 'common/models/Status'
+import Theme from 'common/models/Theme'
 import { Track } from 'common/models/Track'
 import { User } from 'common/models/User'
 import { formatCount } from 'common/utils/formatUtil'
@@ -19,7 +20,6 @@ import Page from 'components/page/Page'
 import TableOptionsButton from 'components/tracks-table/TableOptionsButton'
 import TracksTable, { alphaSortFn } from 'components/tracks-table/TracksTable'
 import useTabs, { useTabRecalculator } from 'hooks/useTabs/useTabs'
-import Theme from 'models/Theme'
 import { getTheme } from 'store/application/ui/theme/selectors'
 import { AppState } from 'store/types'
 import lazyWithPreload from 'utils/lazyWithPreload'
