+++ src/pages/audio-rewards-page/components/modals/TrendingRewards.tsx
@@ -6,6 +6,7 @@ import { useDispatch } from 'react-redux'
 import { TwitterTweetEmbed } from 'react-twitter-embed'
 
 import { useModalState } from 'common/hooks/useModalState'
+import Theme from 'common/models/Theme'
 import { StringKeys } from 'common/services/remote-config'
 import { getTrendingRewardsModalType } from 'common/store/pages/audio-rewards/selectors'
 import {
@@ -16,7 +17,6 @@ import LoadingSpinner from 'components/loading-spinner/LoadingSpinner'
 import { useNavigateToPage } from 'hooks/useNavigateToPage'
 import { useRemoteVar } from 'hooks/useRemoteConfig'
 import { useWithMobileStyle } from 'hooks/useWithMobileStyle'
-import Theme from 'models/Theme'
 import { isMobile } from 'utils/clientUtil'
 import { useSelector } from 'utils/reducer'
 import {
