+++ src/pages/collection-page/store/types.ts
@@ -1,10 +1,10 @@
 import { Moment } from 'moment'
 
 import { ID, UID } from 'common/models/Identifiers'
+import { LineupState } from 'common/models/Lineup'
 import { SmartCollectionVariant } from 'common/models/SmartCollectionVariant'
 import Status from 'common/models/Status'
 import { LineupTrack } from 'common/models/Track'
-import { LineupState } from 'models/common/Lineup'
 
 export type CollectionsPageState = {
   collectionId: ID | null
