+++ src/pages/deleted-page/DeletedPageProvider.tsx
@@ -4,9 +4,9 @@ import { push as pushRoute } from 'connected-react-router'
 import { connect } from 'react-redux'
 import { Dispatch } from 'redux'
 
+import { Playable } from 'common/models/Playable'
 import { User } from 'common/models/User'
 import { LineupVariant } from 'components/lineup/types'
-import { Playable } from 'models/Playable'
 import { makeGetLineupMetadatas } from 'store/lineup/selectors'
 import { getPlaying, getBuffering } from 'store/player/selectors'
 import { makeGetCurrent } from 'store/queue/selectors'
