+++ src/pages/deleted-page/components/desktop/DeletedPage.tsx
@@ -8,6 +8,7 @@ import {
 } from 'common/hooks/useImageSize'
 import { ID, PlayableType } from 'common/models/Identifiers'
 import { CoverArtSizes, SquareSizes } from 'common/models/ImageSizes'
+import { Playable } from 'common/models/Playable'
 import { User } from 'common/models/User'
 import { NestedNonNullable } from 'common/utils/typeUtils'
 import ArtistPopover from 'components/artist/ArtistPopover'
@@ -18,7 +19,6 @@ import NavBanner from 'components/nav-banner/NavBanner'
 import Page from 'components/page/Page'
 import StatBanner from 'components/stat-banner/StatBanner'
 import UserBadges from 'components/user-badges/UserBadges'
-import { Playable } from 'models/Playable'
 import { withNullGuard } from 'utils/withNullGuard'
 
 import styles from './DeletedPage.module.css'
