+++ src/pages/deleted-page/components/mobile/DeletedPage.tsx
@@ -8,6 +8,7 @@ import {
 } from 'common/hooks/useImageSize'
 import { ID, PlayableType } from 'common/models/Identifiers'
 import { CoverArtSizes, SquareSizes } from 'common/models/ImageSizes'
+import { Playable } from 'common/models/Playable'
 import { User } from 'common/models/User'
 import { NestedNonNullable } from 'common/utils/typeUtils'
 import ArtistPopover from 'components/artist/ArtistPopover'
@@ -15,7 +16,6 @@ import DynamicImage from 'components/dynamic-image/DynamicImage'
 import Lineup, { LineupWithoutTile } from 'components/lineup/Lineup'
 import MobilePageContainer from 'components/mobile-page-container/MobilePageContainer'
 import UserBadges from 'components/user-badges/UserBadges'
-import { Playable } from 'models/Playable'
 import { withNullGuard } from 'utils/withNullGuard'
 
 import styles from './DeletedPage.module.css'
