+++ src/pages/feed-page/types.ts
@@ -1,7 +1,7 @@
 import FeedFilter from 'common/models/FeedFilter'
 import { ID, UID } from 'common/models/Identifiers'
+import { Lineup } from 'common/models/Lineup'
 import { User } from 'common/models/User'
-import { Lineup } from 'models/common/Lineup'
 
 export interface FeedPageContentProps {
   feedTitle: string
