+++ src/pages/history-page/store/types.ts
@@ -1,7 +1,7 @@
 import { Moment } from 'moment'
 
 import { ID } from 'common/models/Identifiers'
-import { LineupState } from 'models/common/Lineup'
+import { LineupState } from 'common/models/Lineup'
 
 export default interface HistoryPageState {
   tracks: LineupState<{ id: ID; dateListened: Moment }>
