+++ src/pages/profile-page/components/desktop/ProfilePage.tsx
@@ -9,6 +9,7 @@ import { Name } from 'common/models/Analytics'
 import { Collection } from 'common/models/Collection'
 import { ID, UID } from 'common/models/Identifiers'
 import { CoverPhotoSizes, ProfilePictureSizes } from 'common/models/ImageSizes'
+import { LineupState } from 'common/models/Lineup'
 import Status from 'common/models/Status'
 import { User } from 'common/models/User'
 import Card from 'components/card/desktop/Card'
@@ -25,7 +26,6 @@ import UploadChip from 'components/upload/UploadChip'
 import { useSelectTierInfo } from 'components/user-badges/hooks'
 import { badgeTiers } from 'components/user-badges/utils'
 import useTabs, { useTabRecalculator } from 'hooks/useTabs/useTabs'
-import { LineupState } from 'models/common/Lineup'
 import { MIN_COLLECTIBLES_TIER } from 'pages/profile-page/ProfilePageProvider'
 import EmptyTab from 'pages/profile-page/components/EmptyTab'
 import { feedActions } from 'pages/profile-page/store/lineups/feed/actions'
