+++ src/pages/profile-page/components/mobile/ProfilePage.tsx
@@ -10,6 +10,7 @@ import { ReactComponent as IconReposts } from 'assets/img/iconRepost.svg'
 import { Collection } from 'common/models/Collection'
 import { ID, UID } from 'common/models/Identifiers'
 import { CoverPhotoSizes, ProfilePictureSizes } from 'common/models/ImageSizes'
+import { LineupState } from 'common/models/Lineup'
 import Status from 'common/models/Status'
 import { User } from 'common/models/User'
 import Card from 'components/card/mobile/Card'
@@ -30,7 +31,6 @@ import { useSelectTierInfo } from 'components/user-badges/hooks'
 import { badgeTiers } from 'components/user-badges/utils'
 import useAsyncPoll from 'hooks/useAsyncPoll'
 import useTabs from 'hooks/useTabs/useTabs'
-import { LineupState } from 'models/common/Lineup'
 import { MIN_COLLECTIBLES_TIER } from 'pages/profile-page/ProfilePageProvider'
 import { feedActions } from 'pages/profile-page/store/lineups/feed/actions'
 import { tracksActions } from 'pages/profile-page/store/lineups/tracks/actions'
