+++ src/pages/profile-page/store/types.ts
@@ -1,7 +1,7 @@
 import { ID, UID } from 'common/models/Identifiers'
+import { LineupState } from 'common/models/Lineup'
 import Status from 'common/models/Status'
 import { User } from 'common/models/User'
-import { LineupState } from 'models/common/Lineup'
 
 export enum FollowType {
   FOLLOWERS = 'followers',
