+++ src/pages/saved-page/store/types.ts
@@ -3,8 +3,8 @@ import { Moment } from 'moment'
 import { Collection } from 'common/models/Collection'
 import { Favorite } from 'common/models/Favorite'
 import { UID, ID } from 'common/models/Identifiers'
+import { LineupState } from 'common/models/Lineup'
 import { LineupTrack } from 'common/models/Track'
-import { LineupState } from 'models/common/Lineup'
 
 export default interface SavesPageState {
   localSaves: { [id: number]: UID }
