+++ src/pages/search-page/components/mobile/SearchPageContent.tsx
@@ -13,6 +13,7 @@ import { ReactComponent as IconUser } from 'assets/img/iconUser.svg'
 import { Name } from 'common/models/Analytics'
 import { UserCollection } from 'common/models/Collection'
 import { UID } from 'common/models/Identifiers'
+import { LineupState } from 'common/models/Lineup'
 import Status from 'common/models/Status'
 import { User } from 'common/models/User'
 import Card from 'components/card/mobile/Card'
@@ -27,7 +28,6 @@ import NavContext, {
   RightPreset
 } from 'components/nav/store/context'
 import useTabs from 'hooks/useTabs/useTabs'
-import { LineupState } from 'models/common/Lineup'
 import { tracksActions } from 'pages/search-page/store/lineups/tracks/actions'
 import { make, useRecord } from 'store/analytics/actions'
 import { getLocationPathname } from 'store/routing/selectors'
