+++ src/pages/settings-page/components/desktop/SettingsPage.tsx
@@ -14,6 +14,8 @@ import cn from 'classnames'
 import audiusIcon from 'assets/img/audiusIcon.png'
 import { ID } from 'common/models/Identifiers'
 import { ProfilePictureSizes } from 'common/models/ImageSizes'
+import { OS } from 'common/models/OS'
+import Theme from 'common/models/Theme'
 import { InstagramProfile } from 'common/store/account/reducer'
 import { ChangePasswordModal } from 'components/change-password/ChangePasswordModal'
 import ConfirmationBox from 'components/confirmation-box/ConfirmationBox'
@@ -23,8 +25,6 @@ import Page from 'components/page/Page'
 import SelectedServices from 'components/service-selection/SelectedServices'
 import Toast from 'components/toast/Toast'
 import { ComponentPlacement } from 'components/types'
-import { OS } from 'models/OS'
-import Theme from 'models/Theme'
 import AudiusBackend from 'services/AudiusBackend'
 import DownloadApp from 'services/download-app/DownloadApp'
 import { isMobile, isElectron, getOS } from 'utils/clientUtil'
