+++ src/pages/settings-page/components/mobile/SettingsPage.tsx
@@ -6,6 +6,7 @@ import horizontalLogo from 'assets/img/settingsPageLogo.png'
 import { useUserProfilePicture } from 'common/hooks/useImageSize'
 import { ID } from 'common/models/Identifiers'
 import { SquareSizes, ProfilePictureSizes } from 'common/models/ImageSizes'
+import Theme from 'common/models/Theme'
 import { InstagramProfile, TwitterProfile } from 'common/store/account/reducer'
 import TabSlider from 'components/data-entry/TabSlider'
 import DynamicImage from 'components/dynamic-image/DynamicImage'
@@ -15,7 +16,6 @@ import Row from 'components/groupable-list/Row'
 import NavContext, { LeftPreset } from 'components/nav/store/context'
 import Page from 'components/page/Page'
 import useScrollToTop from 'hooks/useScrollToTop'
-import Theme from 'models/Theme'
 import { getIsIOS } from 'utils/browser'
 import {
   ACCOUNT_SETTINGS_PAGE,
