+++ src/pages/track-page/components/desktop/TrackPage.tsx
@@ -3,6 +3,7 @@ import React from 'react'
 import cn from 'classnames'
 
 import { ID, CID } from 'common/models/Identifiers'
+import { LineupState } from 'common/models/Lineup'
 import { Track } from 'common/models/Track'
 import { User } from 'common/models/User'
 import CoverPhoto from 'components/cover-photo/CoverPhoto'
@@ -14,7 +15,6 @@ import SectionButton from 'components/section-button/SectionButton'
 import StatBanner from 'components/stat-banner/StatBanner'
 import GiantTrackTile from 'components/track/GiantTrackTile'
 import { TrackTileSize } from 'components/track/types'
-import { LineupState } from 'models/common/Lineup'
 import { tracksActions } from 'pages/track-page/store/lineups/tracks/actions'
 import { getTrackDefaults, emptyStringGuard } from 'pages/track-page/utils'
 import { QueueItem } from 'store/queue/types'
