+++ src/pages/trending-page/store/types.ts
@@ -1,6 +1,6 @@
 import { ID } from 'common/models/Identifiers'
+import { LineupState } from 'common/models/Lineup'
 import TimeRange from 'common/models/TimeRange'
-import { LineupState } from 'models/common/Lineup'
 
 export default interface TrendingPageState {
   trendingWeek: LineupState<{ id: ID }>
