+++ src/pages/trending-page/types.ts
@@ -1,8 +1,8 @@
 import { ID, UID } from 'common/models/Identifiers'
+import { Lineup } from 'common/models/Lineup'
 import TimeRange from 'common/models/TimeRange'
 import { Track } from 'common/models/Track'
 import { User } from 'common/models/User'
-import { Lineup } from 'models/common/Lineup'
 
 type ExtraTrendingLineupProps = {}
