+++ src/store/application/ui/theme/types.ts
@@ -1,4 +1,4 @@
-import Theme from 'models/Theme'
+import Theme from 'common/models/Theme'
 
 export interface ThemeState {
   theme: Theme | null
